sealed Console, impl/if separation
