+++ src/java.base/share/classes/java/io/Console.java
@@ -33,8 +33,6 @@ import jdk.internal.access.JavaIOAccess;
 import jdk.internal.access.SharedSecrets;
 import jdk.internal.io.JdkConsoleProvider;
 import jdk.internal.util.StaticProperty;
-import sun.nio.cs.StreamDecoder;
-import sun.nio.cs.StreamEncoder;
 import sun.security.action.GetPropertyAction;
 
 /**
@@ -82,7 +80,7 @@ import sun.security.action.GetPropertyAction;
  * manually zero the returned character array after processing to minimize the
  * lifetime of sensitive data in memory.
  *
- * <blockquote><pre>{@code
+ * {@snippet lang=java :
  * Console cons;
  * char[] passwd;
  * if ((cons = System.console()) != null &&
@@ -90,13 +88,14 @@ import sun.security.action.GetPropertyAction;
  *     ...
  *     java.util.Arrays.fill(passwd, ' ');
  * }
- * }</pre></blockquote>
+ * }
  *
  * @author  Xueming Shen
  * @since   1.6
+ * @sealedGraph
  */
 
-public class Console implements Flushable
+public sealed class Console implements Flushable permits ConsoleImpl, ProxyingConsole
 {
    /**
     * Retrieves the unique {@link java.io.PrintWriter PrintWriter} object
@@ -105,7 +104,8 @@ public class Console implements Flushable
     * @return  The printwriter associated with this console
     */
     public PrintWriter writer() {
-        return pw;
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
 
    /**
@@ -140,7 +140,8 @@ public class Console implements Flushable
     * @return  The reader associated with this console
     */
     public Reader reader() {
-        return reader;
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
 
    /**
@@ -174,8 +175,8 @@ public class Console implements Flushable
     * @return  This console
     */
     public Console format(String fmt, Object ...args) {
-        formatter.format(fmt, args).flush();
-        return this;
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
 
    /**
@@ -214,7 +215,8 @@ public class Console implements Flushable
     * @return  This console
     */
     public Console printf(String format, Object ... args) {
-        return format(format, args);
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
 
    /**
@@ -249,21 +251,8 @@ public class Console implements Flushable
     *          if an end of stream has been reached.
     */
     public String readLine(String fmt, Object ... args) {
-        String line = null;
-        synchronized (writeLock) {
-            synchronized(readLock) {
-                if (!fmt.isEmpty())
-                    pw.format(fmt, args);
-                try {
-                    char[] ca = readline(false);
-                    if (ca != null)
-                        line = new String(ca);
-                } catch (IOException x) {
-                    throw new IOError(x);
-                }
-            }
-        }
-        return line;
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
 
    /**
@@ -277,7 +266,8 @@ public class Console implements Flushable
     *          if an end of stream has been reached.
     */
     public String readLine() {
-        return readLine("");
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
 
    /**
@@ -313,64 +303,8 @@ public class Console implements Flushable
     *          or {@code null} if an end of stream has been reached.
     */
     public char[] readPassword(String fmt, Object ... args) {
-        char[] passwd = null;
-        synchronized (writeLock) {
-            synchronized(readLock) {
-                installShutdownHook();
-                try {
-                    restoreEcho = echo(false);
-                } catch (IOException x) {
-                    throw new IOError(x);
-                }
-                IOError ioe = null;
-                try {
-                    if (!fmt.isEmpty())
-                        pw.format(fmt, args);
-                    passwd = readline(true);
-                } catch (IOException x) {
-                    ioe = new IOError(x);
-                } finally {
-                    try {
-                        if (restoreEcho)
-                            restoreEcho = echo(true);
-                    } catch (IOException x) {
-                        if (ioe == null)
-                            ioe = new IOError(x);
-                        else
-                            ioe.addSuppressed(x);
-                    }
-                    if (ioe != null)
-                        throw ioe;
-                }
-                pw.println();
-            }
-        }
-        return passwd;
-    }
-
-    private void installShutdownHook() {
-        if (shutdownHookInstalled)
-            return;
-        try {
-            // Add a shutdown hook to restore console's echo state should
-            // it be necessary.
-            SharedSecrets.getJavaLangAccess()
-                .registerShutdownHook(0 /* shutdown hook invocation order */,
-                    false /* only register if shutdown is not in progress */,
-                    new Runnable() {
-                        public void run() {
-                            try {
-                                if (restoreEcho) {
-                                    echo(true);
-                                }
-                            } catch (IOException x) { }
-                        }
-                    });
-        } catch (IllegalStateException e) {
-            // shutdown is already in progress and readPassword is first used
-            // by a shutdown hook
-        }
-        shutdownHookInstalled = true;
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
 
    /**
@@ -384,7 +318,8 @@ public class Console implements Flushable
     *          or {@code null} if an end of stream has been reached.
     */
     public char[] readPassword() {
-        return readPassword("");
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
 
     /**
@@ -392,9 +327,9 @@ public class Console implements Flushable
      * immediately .
      */
     public void flush() {
-        pw.flush();
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
-
 
     /**
      * Returns the {@link java.nio.charset.Charset Charset} object used for
@@ -410,20 +345,10 @@ public class Console implements Flushable
      * @since 17
      */
     public Charset charset() {
-        assert CHARSET != null : "charset() should not return null";
-        return CHARSET;
+        throw new UnsupportedOperationException(
+                "Console class itself does not provide implementation");
     }
 
-    private Object readLock;
-    private Object writeLock;
-    private Reader reader;
-    private Writer out;
-    private PrintWriter pw;
-    private Formatter formatter;
-    private char[] rcb;
-    private boolean restoreEcho;
-    private boolean shutdownHookInstalled;
-    private static native String encoding();
     /*
      * Sets the console echo status to {@code on} and returns the previous
      * console on/off status.
@@ -431,150 +356,10 @@ public class Console implements Flushable
      *              {@code false} for echo off
      * @return true if the previous console echo status is on
      */
-    private static native boolean echo(boolean on) throws IOException;
+    static native boolean echo(boolean on) throws IOException;
+    private static native String encoding();
+    static final Charset CHARSET;
 
-    private char[] readline(boolean zeroOut) throws IOException {
-        int len = reader.read(rcb, 0, rcb.length);
-        if (len < 0)
-            return null;  //EOL
-        if (rcb[len-1] == '\r')
-            len--;        //remove CR at end;
-        else if (rcb[len-1] == '\n') {
-            len--;        //remove LF at end;
-            if (len > 0 && rcb[len-1] == '\r')
-                len--;    //remove the CR, if there is one
-        }
-        char[] b = new char[len];
-        if (len > 0) {
-            System.arraycopy(rcb, 0, b, 0, len);
-            if (zeroOut) {
-                Arrays.fill(rcb, 0, len, ' ');
-            }
-        }
-        return b;
-    }
-
-    private char[] grow() {
-        assert Thread.holdsLock(readLock);
-        char[] t = new char[rcb.length * 2];
-        System.arraycopy(rcb, 0, t, 0, rcb.length);
-        rcb = t;
-        return rcb;
-    }
-
-    class LineReader extends Reader {
-        private Reader in;
-        private char[] cb;
-        private int nChars, nextChar;
-        boolean leftoverLF;
-        LineReader(Reader in) {
-            this.in = in;
-            cb = new char[1024];
-            nextChar = nChars = 0;
-            leftoverLF = false;
-        }
-        public void close () {}
-        public boolean ready() throws IOException {
-            //in.ready synchronizes on readLock already
-            return in.ready();
-        }
-
-        public int read(char[] cbuf, int offset, int length)
-            throws IOException
-        {
-            int off = offset;
-            int end = offset + length;
-            if (offset < 0 || offset > cbuf.length || length < 0 ||
-                end < 0 || end > cbuf.length) {
-                throw new IndexOutOfBoundsException();
-            }
-            synchronized(readLock) {
-                boolean eof = false;
-                char c = 0;
-                for (;;) {
-                    if (nextChar >= nChars) {   //fill
-                        int n = 0;
-                        do {
-                            n = in.read(cb, 0, cb.length);
-                        } while (n == 0);
-                        if (n > 0) {
-                            nChars = n;
-                            nextChar = 0;
-                            if (n < cb.length &&
-                                cb[n-1] != '\n' && cb[n-1] != '\r') {
-                                /*
-                                 * we're in canonical mode so each "fill" should
-                                 * come back with an eol. if there no lf or nl at
-                                 * the end of returned bytes we reached an eof.
-                                 */
-                                eof = true;
-                            }
-                        } else { /*EOF*/
-                            if (off - offset == 0)
-                                return -1;
-                            return off - offset;
-                        }
-                    }
-                    if (leftoverLF && cbuf == rcb && cb[nextChar] == '\n') {
-                        /*
-                         * if invoked by our readline, skip the leftover, otherwise
-                         * return the LF.
-                         */
-                        nextChar++;
-                    }
-                    leftoverLF = false;
-                    while (nextChar < nChars) {
-                        c = cbuf[off++] = cb[nextChar];
-                        cb[nextChar++] = 0;
-                        if (c == '\n') {
-                            return off - offset;
-                        } else if (c == '\r') {
-                            if (off == end) {
-                                /* no space left even the next is LF, so return
-                                 * whatever we have if the invoker is not our
-                                 * readLine()
-                                 */
-                                if (cbuf == rcb) {
-                                    cbuf = grow();
-                                    end = cbuf.length;
-                                } else {
-                                    leftoverLF = true;
-                                    return off - offset;
-                                }
-                            }
-                            if (nextChar == nChars && in.ready()) {
-                                /*
-                                 * we have a CR and we reached the end of
-                                 * the read in buffer, fill to make sure we
-                                 * don't miss a LF, if there is one, it's possible
-                                 * that it got cut off during last round reading
-                                 * simply because the read in buffer was full.
-                                 */
-                                nChars = in.read(cb, 0, cb.length);
-                                nextChar = 0;
-                            }
-                            if (nextChar < nChars && cb[nextChar] == '\n') {
-                                cbuf[off++] = '\n';
-                                nextChar++;
-                            }
-                            return off - offset;
-                        } else if (off == end) {
-                           if (cbuf == rcb) {
-                                cbuf = grow();
-                                end = cbuf.length;
-                           } else {
-                               return off - offset;
-                           }
-                        }
-                    }
-                    if (eof)
-                        return off - offset;
-                }
-            }
-        }
-    }
-
-    private static final Charset CHARSET;
     static {
         Charset cs = null;
         boolean istty = istty();
@@ -619,31 +404,17 @@ public class Console implements Flushable
                         .filter(Objects::nonNull)
                         .findAny()
                         .map(jc -> (Console) new ProxyingConsole(jc))
-                        .orElse(istty ? new Console() : null);
+                        .orElse(istty ? new ConsoleImpl() : null);
             };
             return AccessController.doPrivileged(pa);
         } catch (ServiceConfigurationError ignore) {
             // default to built-in Console
-            return istty ? new Console() : null;
+            return istty ? new ConsoleImpl() : null;
         }
     }
 
     private static final Console cons;
     private static native boolean istty();
 
-    Console() {
-        readLock = new Object();
-        writeLock = new Object();
-        out = StreamEncoder.forOutputStreamWriter(
-                  new FileOutputStream(FileDescriptor.out),
-                  writeLock,
-                  CHARSET);
-        pw = new PrintWriter(out, true) { public void close() {} };
-        formatter = new Formatter(out);
-        reader = new LineReader(StreamDecoder.forInputStreamReader(
-                     new FileInputStream(FileDescriptor.in),
-                     readLock,
-                     CHARSET));
-        rcb = new char[1024];
-    }
+    Console() {}
 }
